+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * bump github.com/briandowns/spinner from 1.12.0 to 1.13.0
 * fix(backup-download) Now download the last successful backup [#663](https://github.com/Scalingo/cli/pull/663)
 * fix(review-app-url) add app url in the tablewriter [#664](https://github.com/Scalingo/cli/pull/664)
+* fix(backups-download) print error on stderr if the backup status is not 'done' [#665](https://github.com/Scalingo/cli/pull/665)
 
 ### 1.20.2